Update load triggers branch for IE
The solution inspired by http://dean.edwards.name/weblog/2006/06/again/ 
doesn't work correctly. The script insertion causes an "do you want to display the insecure content?" warning because the URL javascript:void(0) is not considered to be secure. 
However we have been able to replace it with a simple reference to document.readyState so the script is not necessary at all.